labs/03-langchain/demo-06-non-blocking-chat-with-llm/pyproject.toml
[project]
name = "demo-6-non-blocking-chat-with-llm"
version = "0.1.0"
description = "LangChain + Gemini with Async Operations"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain==1.0.0",
    "langchain-openai==1.0.0",
    "langchain-core==1.0.1",
    "python-dotenv==1.2.1",
    "fastapi==0.121.0",
    "uvicorn==0.38.0",
    "pydantic==2.12.4",
]
